requirements.txt
python>=3.10
python-multipart

# Serveur web et Framework
fastapi>=0.115
uvicorn>=0.34

# Base de données
sqlalchemy>=2.0.40
aiosqlite>=0.21
aiomysql>=0.2

# Authentification
python-jose>=3.4
argon2-cffi

# Pydantic v2
pydantic>=2.11
pydantic-settings>=2.9.1

# Dotenv
python-dotenv>=1.1

annotated-types>=0.7.0

# Tests
pytest
pytest-asyncio
httpx
